refactor: improve PicoLTE device monitoring script with error handling and function renames
* refactor: add error handling to all AT command interactions
* rename: check_cat_m1_and_network_type -> check_network_type
* rename: check_signal_quality_first -> check_signal_quality
* move: relocate monitoring script from tools/ to example/__basic__/
* improve: ensure continuous execution despite individual check failures
